Refactor to allow access to the context functions
Vim allows you to have the backtick-enclosed text expanded as a Vim
expression i(nstead of as an external command) by putting an equal sign
right after the first backtick. For example:

    :e `=tempname()`

See:

    :h `=

for more.

This got me thinking: what if a user wants to perform some operation on
the files captured by a given 'file context' (e.g. treeish) that I have
not yet anticipated? It would be super nice to have the functions which
actually get these files exposed as Vimscript functions so that they can
be used with backtick expansion. For example:

    :args `=TreeishFiles("<commit-hash>", "<pathspec>")`

is equivalent to the helper command:

    :ArgsTreeish <commit-hash> <pathspec>

This commit refactors the plugin to support this use case. This has had
the nice side effect of making the code a little cleaner as well.