Add 'Tracked' context
Allows you to manipulate the arglist based on files that are _tracked_
in a given Git treeish.

Motivation: I was working on a project where, as part of the build
process, a load of ephemeral files would be spat out at various points
in the working tree. I needed a way to set the arglist to all of the
files beneath a particular subdirectory, excluding anything from the
build process. This could have been achieved with something like:

    :args subdirectory/**/*
    :ArgDeleteUntracked subdirectory

... but it's far nicer to be able to do it with just one command.

Now you can just do:

    :ArgsTracked subdirectory